+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Notes are displayed like in chat messaging apps (bubbles)
 - Update vuetify to 2.1.3
 
 ### Removed