+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Schemas
+- Schemas for collections
 
 ### Changed
 - Update occitan translation, by Quenty31